# Conflicts:
#   templates/series/_preview.html.twig

+++ src/Controller/RateController.php
@@ -2,35 +2,36 @@
 
 namespace App\Controller;
 
-use DateTime;
+use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
+use Symfony\Component\HttpFoundation\Response;
+use Symfony\Component\Routing\Annotation\Route;
+use Symfony\Component\HttpFoundation\Request;
+use Symfony\Component\Form\FormInterface;
+use App\Form\RateFormType;
+use Doctrine\ORM\EntityManagerInterface;
 use App\Entity\Rating;
 use App\Entity\Series;
 use App\Repository\SeriesRepository;
 use App\Repository\RatingRepository;
-use App\Form\RateFormType;
-use Symfony\Component\Form\FormInterface;
-use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
-use Doctrine\ORM\EntityManagerInterface;
-use Symfony\Component\HttpFoundation\Request;
-use Symfony\Component\HttpFoundation\Response;
-use Symfony\Component\Routing\Annotation\Route;
+use DateTime;
 
 /**
  * @Route("/user/rate")
  */
 class RateController extends AbstractController
 {
+    // The paths of the pages to render
+    private static $newRatePage = 'home/index.html.twig';
+    private static $seeRatePage = 'home/about.html.twig';
+
     /**
-     * @Route("/add/{id}", name="add_series_rate", methods={"GET"})
+     * @Route("/add/{id}", name="addSeriesRate")
      */
-    public function add_rate(Series $serie, Request $request, SeriesRepository $repository, EntityManagerInterface $entityManager): Response
+    public function addRate(Series $serie, Request $request, EntityManagerInterface $entityManager): Response
     {
-        // Verify that the user is loged in and not an admin
-        if ($this->getUser() == null) {
-            return $this->redirectToRoute('user_login');
-        }
-        if ($this->getUser()->getAdmin()) {
-            return $this->redirectToRoute('home');
+        // Verify that the user is connected and is a normal user
+        if ($isNotNormalUser = $this->verifyNormalUser()) {
+            return $isNotNormalUser;
         }
 
         $rate = new Rating();
@@ -39,27 +40,24 @@ class RateController extends AbstractController
 
         // If form was submited and valid, then add the rating and go back to serie's page
         if ($form->isSubmitted() && $form->isValid()) {
-            return $this->save_rate($serie, $form, $entityManager);
+            return $this->saveRate($serie, $form, $entityManager);
         }
 
         // Else render the form
-        return $this->render('user/rate.html.twig', [
+        return $this->render(RateController::$newRatePage, [
             'serie' => $serie,
             'form' => $form->createView()
         ]);
     }
 
     /**
-     * @Route("/modify/{id}", name="modify_series_rate", methods={"GET"})
+     * @Route("/modify/{id}", name="modifySeriesRate")
      */
-    public function modify_rate(Series $serie, Request $request, SeriesRepository $repository, EntityManagerInterface $entityManager): Response
+    public function modifySeriesRate(Series $serie, Request $request, SeriesRepository $repository, EntityManagerInterface $entityManager): Response
     {
-        // Verify that the user is loged in and not an admin
-        if ($this->getUser() == null) {
-            return $this->redirectToRoute('user_login');
-        }
-        if ($this->getUser()->getAdmin()) {
-            return $this->redirectToRoute('home');
+        // Verify that the user is connected and is a normal user
+        if ($isNotNormalUser = $this->verifyNormalUser()) {
+            return $isNotNormalUser;
         }
 
         $rating = $serie->getRatingByUser($this->getUser());
@@ -68,43 +66,45 @@ class RateController extends AbstractController
 
         // If form was submited and valid, then modift the rating and go back to the serie's page
         if ($form->isSubmitted() && $form->isValid()) {
-            return $this->save_rate_modification($rating, $form, $entityManager);
+            return $this->saveRateModification($rating, $form, $entityManager);
         }
 
         // Else render the form
-        return $this->render('user/rate.html.twig', [
+        return $this->render(RateController::$newRatePage, [
             'serie' => $serie,
             'form' => $form->createView()
         ]);
     }
 
     /**
-     * @Route("/delete/{id}", name="delete_series_rate", methods={"GET"})
+     * @Route("/delete/{id}", name="deleteSeriesRate")
      */
-    public function delete_rate(RatingRepository $ratingRepository, int $id, EntityManagerInterface $entityManager): Response
+    public function deleteSeriesRate(RatingRepository $ratingRepository, int $id, Request $request, EntityManagerInterface $entityManager): Response
     {
+        // Verify that the user is connected and is a normal user
+        if ($isNotNormalUser = $this->verifyNormalUser()) {
+            return $isNotNormalUser;
+        }
+
         $rate = $ratingRepository->find($id);
         $entityManager->remove($rate); 
         $entityManager->flush();       
 
-        return $this->redirectToRoute('comments_moderation');
+        return $this->redirect($request->headers->get('referer'));
     }
 
 
     /**
-     * @Route("/see/{id}", name="see_rate")
+     * @Route("/see/{id}", name="seeRate")
      */
-    public function see_rate(Series $serie): Response
+    public function seeRate(Series $serie): Response
     {
-        // Verify that the user is loged in and not an admin
-        if ($this->getUser() == null) {
-            return $this->redirectToRoute('user_login');
-        }
-        if ($this->getUser()->getAdmin()) {
-            return $this->redirectToRoute('home');
+        // Verify that the user is connected and is a normal user
+        if ($isNotNormalUser = $this->verifyNormalUser()) {
+            return $isNotNormalUser;
         }
         
-        return $this->render('user/see_rate.html.twig', [
+        return $this->render(RateController::$seeRatePage, [
             'serie' => $serie,
             'rating' => $serie->getRatingByUser($this->getUser()),
         ]);
@@ -113,7 +113,7 @@ class RateController extends AbstractController
     /**
      * Save the given serie's rating to the database.
      */
-    private function save_rate(Series $serie, FormInterface $form, EntityManagerInterface $entityManager): Response
+    private function saveRate(Series $serie, FormInterface $form, EntityManagerInterface $entityManager): Response
     {
         // Creation & initialisation of the rating
         $rate = new Rating();
@@ -126,13 +126,13 @@ class RateController extends AbstractController
         $entityManager->persist($rate); // save rating
         $entityManager->flush();    // Update the changes made in the databse
 
-        return $this->redirectToRoute('series_presentation', ['id' => $serie->getId()]);
+        return $this->redirectToRoute('seriesPresentation', ['id' => $serie->getId()]);
     }
 
     /**
      * Modify the given serie's rating of the current user.
      */
-    private function save_rate_modification(Rating $rate, FormInterface $form, EntityManagerInterface $entityManager): Response
+    private function saveRateModification(Rating $rate, FormInterface $form, EntityManagerInterface $entityManager): Response
     {
         // Creation & initialisation of the rating
         $rate->setValue((int)$form->get('value')->getData());
@@ -142,6 +142,24 @@ class RateController extends AbstractController
         $entityManager->persist($rate); // save rating
         $entityManager->flush();        // Update the changes made in the databse
 
-        return $this->redirectToRoute('series_presentation', ['id' => $rate->getSeries()->getId()]);
+        return $this->redirectToRoute('seriesPresentation', ['id' => $rate->getSeries()->getId()]);
+    }
+
+    /**
+     * Verify that the user is connected and is not an admin.
+     * 
+     * @return Response to login page if the user is not connected
+     * @return Response to home page if the user connected is an admin
+     * @return null if the user connected is not an admin
+     */
+    private function verifyNormalUser(): ?Response
+    {
+        if ($this->getUser() == null) {
+            return $this->redirectToRoute('userLogin');
+        }
+        if ($this->getUser()->getAdmin()) {
+            return $this->redirectToRoute('home');
+        }
+        return null;
     }
 }
